+++ frontend/src/app/(common)/vocs/_components/VocDetailActions.tsx
@@ -52,7 +52,7 @@ export function VocDetailActions({ vocId, status }: { vocId: number; status: str
       <LoginRequiredModal isOpen={showLoginModal} onClose={() => setShowLoginModal(false)} />
       <div className="mt-10 flex flex-wrap items-center justify-end gap-3 border-t border-primary/10 pt-10">
         <Link
-          href={`/profile/vocs/${vocId}/edit`}
+          href={`/vocs/${vocId}/edit`}
           className="inline-flex items-center gap-2 rounded-full border border-secondary bg-secondary/15 px-5 py-2.5 text-xs font-black uppercase tracking-[0.24em] text-secondary transition-transform hover:-translate-y-0.5"
         >
           <Pencil className="size-4" aria-hidden />
