+++ frontend/src/app/(common)/vocs/new/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
 
 export default function LegacyVocNewRedirect() {
-  redirect("/profile/vocs");
+  redirect("/vocs");
 }
